refactor host_buffer/functions
