+++ ttnn/core/tensor/host_buffer/functions.cpp
@@ -3,6 +3,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include <tt_stl/reflection.hpp>
+#include <type_traits>
 #include <vector>
 
 #include <ttnn/tensor/host_buffer/functions.hpp>
@@ -10,6 +11,36 @@
 #include <ttnn/tensor/tensor_utils.hpp>
 
 namespace tt::tt_metal::host_buffer {
+
+namespace {
+namespace CMAKE_UNIQUE_NAMESPACE {
+
+template <typename T>
+void validate_datatype(const Tensor& tensor) {
+    using BaseType = std::remove_cvref_t<T>;
+    if constexpr (std::is_same_v<BaseType, uint32_t>) {
+        TT_FATAL(
+            tensor.dtype() == DataType::UINT32 or tensor.dtype() == DataType::BFLOAT8_B or
+                tensor.dtype() == DataType::BFLOAT4_B,
+            "Incorrect data type {}",
+            tensor.dtype());
+    } else if constexpr (std::is_same_v<BaseType, int32_t>) {
+        TT_FATAL(tensor.dtype() == DataType::INT32, "Incorrect data type {}", tensor.dtype());
+    } else if constexpr (std::is_same_v<BaseType, float>) {
+        TT_FATAL(tensor.dtype() == DataType::FLOAT32, "Incorrect data type {}", tensor.dtype());
+    } else if constexpr (std::is_same_v<BaseType, bfloat16>) {
+        TT_FATAL(tensor.dtype() == DataType::BFLOAT16, "Incorrect data type {}", tensor.dtype());
+    } else if constexpr (std::is_same_v<BaseType, uint16_t>) {
+        TT_FATAL(tensor.dtype() == DataType::UINT16, "Incorrect data type {}", tensor.dtype());
+    } else if constexpr (std::is_same_v<BaseType, uint8_t>) {
+        TT_FATAL(tensor.dtype() == DataType::UINT8, "Incorrect data type {}", tensor.dtype());
+    } else {
+        static_assert(sizeof(BaseType) == 0, "Unsupported DataType");
+    }
+}
+
+}  // namespace CMAKE_UNIQUE_NAMESPACE
+}  // namespace
 
 HostBuffer get_host_buffer(const Tensor& tensor) {
     TT_FATAL(is_cpu_tensor(tensor), "Tensor must have HostStorage");
@@ -22,4 +53,49 @@ HostBuffer get_host_buffer(const Tensor& tensor) {
         storage.buffer().shape());
     return buffers.front();
 }
+
+template <typename T>
+ttsl::Span<const T> get_as(const HostBuffer& buffer) {
+    return buffer.view_as<T>();
+}
+
+template <typename T>
+ttsl::Span<T> get_as(HostBuffer& buffer) {
+    return buffer.view_as<T>();
+}
+
+template <typename T>
+ttsl::Span<const T> get_as(const Tensor& tensor) {
+    CMAKE_UNIQUE_NAMESPACE::validate_datatype<T>(tensor);
+    HostBuffer buffer = get_host_buffer(tensor);
+    return buffer.template view_as<T>();
+}
+
+template <typename T>
+ttsl::Span<T> get_as(Tensor& tensor) {
+    CMAKE_UNIQUE_NAMESPACE::validate_datatype<T>(tensor);
+    HostBuffer buffer = get_host_buffer(tensor);
+    return buffer.template view_as<T>();
+}
+
+// Explicit template instantiations
+#define INSTANTIATE_HOST_BUFFER_FUNCTIONS(T)                         \
+    template ttsl::Span<const T> get_as<T>(const HostBuffer&);       \
+    template ttsl::Span<const T> get_as<const T>(const HostBuffer&); \
+    template ttsl::Span<T> get_as<T>(HostBuffer&);                   \
+    template ttsl::Span<const T> get_as<const T>(HostBuffer&);       \
+    template ttsl::Span<const T> get_as<T>(const Tensor&);           \
+    template ttsl::Span<const T> get_as<const T>(const Tensor&);     \
+    template ttsl::Span<T> get_as<T>(Tensor&);                       \
+    template ttsl::Span<const T> get_as<const T>(Tensor&);
+
+INSTANTIATE_HOST_BUFFER_FUNCTIONS(uint32_t)
+INSTANTIATE_HOST_BUFFER_FUNCTIONS(int32_t)
+INSTANTIATE_HOST_BUFFER_FUNCTIONS(float)
+INSTANTIATE_HOST_BUFFER_FUNCTIONS(bfloat16)
+INSTANTIATE_HOST_BUFFER_FUNCTIONS(uint16_t)
+INSTANTIATE_HOST_BUFFER_FUNCTIONS(uint8_t)
+
+#undef INSTANTIATE_HOST_BUFFER_FUNCTIONS
+
 }  // namespace tt::tt_metal::host_buffer
